requirements.txt
"python-socketio[asyncio_client]"
RPi.GPIO
requests
